pyproject.toml
[tool.poetry]
name = "inforce"
version = "0.1.0"
description = ""
authors = ["ahzees <matviy.stets@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
djangorestframework-simplejwt = "^5.2.2"
pytest-dotenv = "^0.5.2"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.0.4"
isort = "^5.12.0"
commitizen = "^2.41.0"
django-dotenv = "^1.4.2"
django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
flakehell = "0.7.1"
flake8 = "3.9.0"
drf-spectacular = "^0.26.2"
pytest-django = "^4.5.2"

[tool.flakehell]
max_line_length = 88

[tool.pytest.ini_options]
pythonpath = "inforce"
DJANGO_SETTINGS_MODULE = "inforce.settings"

[tool.isort]
profile = "black"

[tool.commitizen]
name = "cz_customize"
version = "0.1.0"
tag_format = "v$version"
update_changelog_on_bump = true
version_files = [
    "inforce/__init__.py:version",
    "pyproject.toml:^version"
]

[tool.commitizen.customize]
message_template = "{{change_type}}:{% if show_message %} {{message}}{% endif %}"
example = "feature: this feature enable customize through config file"
schema = "<type>: <body>"
bump_pattern = "^(major|minor|patch|hotfix)"
bump_map = {"major" = "MAJOR", "minor" = "MINOR", "patch" = "PATCH", "hotfix" = "PATCH"}

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = "C0114, C0115, R0903, C0116, C0415, W0614, E1101"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
